added chat-delay - alpha-9.2
