+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "the-boct",
-  "version": "4.0.0-alpha-9",
+  "version": "4.0.0-alpha-9.2",
   "license": "MIT",
   "homepage": "https://the-boct.github.io/",
   "private": true,
